add backup package to sidecar
