+++ cmd/sidecar/handler/backup/handler.go
@@ -1,4 +1,4 @@
-package handler
+package backup
 
 import (
 	"context"
@@ -13,63 +13,36 @@ import (
 	"github.com/percona/percona-server-mysql-operator/pkg/xtrabackup/storage"
 )
 
-type Status struct {
-	isRunning         atomic.Bool
-	currentBackupConf *xb.BackupConfig
-
-	mu sync.Mutex
-}
-
-func (s *Status) TryRunBackup() bool {
-	return s.isRunning.CompareAndSwap(false, true)
-}
-
-func (s *Status) DoneBackup() {
-	s.isRunning.Store(false)
-}
-
-func (s *Status) SetBackupConfig(conf xb.BackupConfig) {
-	s.mu.Lock()
-	s.currentBackupConf = &conf
-	s.mu.Unlock()
-}
-
-func (s *Status) RemoveBackupConfig() {
-	s.mu.Lock()
-	s.currentBackupConf = nil
-	s.mu.Unlock()
-}
-
-func (s *Status) GetBackupConfig() *xb.BackupConfig {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.currentBackupConf
-}
-
-type handlerBackup struct {
-	status Status
+type Handler struct {
+	status status
 
 	newStorageFunc   storage.NewClientFunc
 	getNamespaceFunc func() (string, error)
 	deleteBackupFunc func(ctx context.Context, cfg *xb.BackupConfig, backupName string) error
 }
 
-func NewBackupHandler() http.Handler {
-	return &handlerBackup{
-		newStorageFunc: storage.NewClient,
-		getNamespaceFunc: func() (string, error) {
+func (h *Handler) init() {
+	if h.deleteBackupFunc == nil {
+		h.deleteBackupFunc = deleteBackup
+	}
+	if h.newStorageFunc == nil {
+		h.newStorageFunc = storage.NewClient
+	}
+	if h.getNamespaceFunc == nil {
+		h.getNamespaceFunc = func() (string, error) {
 			ns, err := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
 			if err != nil {
 				return "", errors.Wrap(err, "read namespace file")
 			}
 
 			return string(ns), nil
-		},
-		deleteBackupFunc: deleteBackup,
+		}
 	}
 }
 
-func (h *handlerBackup) ServeHTTP(w http.ResponseWriter, req *http.Request) {
+func (h *Handler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
+	h.init()
+
 	switch req.Method {
 	case http.MethodGet:
 		h.getBackupHandler(w, req)
@@ -82,7 +55,7 @@ func (h *handlerBackup) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 	}
 }
 
-func (h *handlerBackup) backupExists(ctx context.Context, cfg *xb.BackupConfig) (bool, error) {
+func (h *Handler) backupExists(ctx context.Context, cfg *xb.BackupConfig) (bool, error) {
 	opts, err := storage.GetOptionsFromBackupConfig(cfg)
 	if err != nil {
 		return false, errors.Wrap(err, "get options from backup config")
@@ -99,4 +72,37 @@ func (h *handlerBackup) backupExists(ctx context.Context, cfg *xb.BackupConfig) 
 		return false, nil
 	}
 	return true, nil
+}
+
+type status struct {
+	isRunning         atomic.Bool
+	currentBackupConf *xb.BackupConfig
+
+	mu sync.Mutex
+}
+
+func (s *status) tryRunBackup() bool {
+	return s.isRunning.CompareAndSwap(false, true)
+}
+
+func (s *status) doneBackup() {
+	s.isRunning.Store(false)
+}
+
+func (s *status) setBackupConfig(conf xb.BackupConfig) {
+	s.mu.Lock()
+	s.currentBackupConf = &conf
+	s.mu.Unlock()
+}
+
+func (s *status) removeBackupConfig() {
+	s.mu.Lock()
+	s.currentBackupConf = nil
+	s.mu.Unlock()
+}
+
+func (s *status) getBackupConfig() *xb.BackupConfig {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	return s.currentBackupConf
 }
